Select the proper test/spec method to run.
Because of the difference in Spock between the name of the feature e.g. "This is the name of some spec" and the actual method name .e.g $spock_feature_0_0, there's a bit more work involved in the logic to select which method to run. Arquillian passes the actual method name, but the JUnit filtering mechanism used in the Sputnik filter passes in the feature name. This commit makes sure that the correct method is selected.